Course comment voting fix

+++ src/api/schemas/delete.py
@@ -1,19 +1,18 @@
 import graphene
 from django.utils.translation import gettext as _
-from graphene_django.forms.mutation import DjangoModelFormMutation
+from graphene_django.forms.mutation import DjangoFormMutation
 from graphql import ResolveInfo
 from graphql_jwt.decorators import login_required
 
 from api.forms.delete import DeleteObjectForm
 
 
-class DeleteObjectMutation(DjangoModelFormMutation):
+# FIXME: This mutation currently returns all form fields (that are always null) but we only want to return the message.
+class DeleteObjectMutation(DjangoFormMutation):
     message = graphene.String()
 
     class Meta:
         form_class = DeleteObjectForm
-        exclude_fields = ("id",)
-        return_field_name = "message"
 
     @classmethod
     @login_required
@@ -21,6 +20,16 @@ class DeleteObjectMutation(DjangoModelFormMutation):
         cls, form: DeleteObjectForm, info: ResolveInfo
     ) -> "DeleteObjectMutation":
         obj = form.cleaned_data.get("target")
+        if obj.user != info.context.user:
+            return cls(
+                errors=[
+                    {
+                        "field": "__all__",
+                        "messages": [_("You are not the owner of this object!")],
+                    }
+                ]
+            )
+
         obj.delete()
         return cls(message=_("Object deleted successfully!"))
